fix(ci): satisfy the clippy and rustdoc gates in Lints
With the fmt failure out of the way the Lints job reached `cargo clippy` and
failed on five lints: two `unnecessary_cast` (`c_int` is already `i32` on every
supported target), `doc_overindented_list_items`, `manual_filter` on the
'noexpandtab' indent start, and `explicit_counter_loop` in the Virtual Replace
test.

`cargo doc` had never run either. Backtick `<PageUp>`/`<PageDown>`/`<CR>` so
rustdoc stops reading them as HTML, and drop the intra-doc links to the plugin
host's private `CxGuard`/`with_cx` — rustdoc resolves a module's inner docs
where those two are out of scope, so the link can never resolve.

+++ zmax-plugin/src/lib.rs
@@ -61,8 +61,8 @@ pub const INIT_SYMBOL: &[u8] = b"zmax_plugin_init\0";
 /// * `host`   — the host API table (call back into the editor through it).
 /// * `argc`   — number of elements in `argv`.
 /// * `argv`   — NUL-terminated C strings; `argv[0]` is the command name,
-///              `argv[1..]` the arguments. Valid only for the duration of the
-///              call; copy anything you need to keep.
+///   `argv[1..]` the arguments. Valid only for the duration of the call; copy
+///   anything you need to keep.
 ///
 /// Returns the command's exit status (0 = success).
 pub type CommandFn =
@@ -189,7 +189,7 @@ impl Host {
     /// Run a `:` command `line` (without the leading `:`); returns its exit status.
     pub fn eval(&self, line: &str) -> i32 {
         match CString::new(line) {
-            Ok(c) => (self.t().eval)(self.api, c.as_ptr()) as i32,
+            Ok(c) => (self.t().eval)(self.api, c.as_ptr()),
             Err(_) => 1,
         }
     }
